Add 'toggle grid lines' functionality
It works by toggling an additional class that overrides the current
grid value. It will not work by attempting to overwrite the current
grid value within board because the grid items are created dynamically.
If attempting to overwrite the current, all other styles on the grid
such as no. of columns will be removed and replaced by the new grid
value, which is problematic.